refactor(core): restructure atoms module into separate files
Move atoms from single file to directory with individual files per atom:
- atoms/mod.rs: Atom trait and re-exports
- atoms/add_user_message.rs: AddUserMessageAtom
- atoms/call_model.rs: CallModelAtom
- atoms/execute_tool.rs: ExecuteToolAtom

This follows the same pattern as the capabilities module for consistency.
